move prepareImages to helper package
