+++ docs/content/docs/helpers.mdx
@@ -43,7 +43,7 @@ Length and color shorthands return CSS strings. `vw`/`vh` are viewport width/hei
 
 | Helper                           | Import                    | Use                                                                                               |
 | :------------------------------- | :------------------------ | :------------------------------------------------------------------------------------------------ |
-| `prepareImages(options)`         | `takumi-js`               | Fetch a node tree's remote images into `images` entries. See [Images & emoji](/docs/load-images). |
+| `prepareImages(options)`         | `takumi-js/helpers`       | Fetch a node tree's remote images into `images` entries. See [Images & emoji](/docs/load-images). |
 | `extractEmojis(node, emojiType)` | `takumi-js/helpers/emoji` | Replace emoji with CDN `<img>` nodes. See [Emoji](/docs/load-images#emoji).                       |
 
 `emojiType` selects the CDN provider: `twemoji`, `blobmoji`, `noto`, `openmoji`, `fluent`, or `fluentFlat`.
