+++ takumi-js/src/render.ts
@@ -1,154 +1,15 @@
 import type * as napi from "@takumi-rs/core";
 import type * as wasm from "@takumi-rs/wasm";
 import { extractEmojis, type EmojiType } from "@takumi-rs/helpers/emoji";
-import { fetchOk, type FetchOptions } from "@takumi-rs/helpers";
+import { prepareImages, type FetchOptions, type ImageFetchCache } from "@takumi-rs/helpers";
 import { fromJsx, type FromJsxOptions } from "@takumi-rs/helpers/jsx";
 import { getImports } from "./import";
-import type { CSSProperties, ReactNode } from "react";
+import type { ReactNode } from "react";
 import type { Node, ReactElementLike } from "@takumi-rs/helpers";
 import { fromHtml } from "@takumi-rs/helpers/html";
 
 type Renderer = napi.Renderer | wasm.Renderer;
 type ImageLoader = napi.ImageLoader | wasm.ImageLoader;
-
-/**
- * A cache of image fetches keyed by URL. Sharing one across renders deduplicates concurrent
- * requests for the same URL (single-flight) and reuses their bytes. Any object with `Map`-like
- * `get`/`set`/`delete` works, so LRU/TTL policies can be plugged in.
- */
-export interface ImageFetchCache {
-  get(url: string): Promise<ArrayBuffer> | undefined;
-  set(url: string, data: Promise<ArrayBuffer>): unknown;
-  delete(url: string): unknown;
-}
-
-const defaultFetchTimeout = 5000;
-const cssUrlPattern = /url\(\s*(['"]?)(.*?)\1\s*\)/g;
-
-function isRemoteUrl(value: string): boolean {
-  return value.startsWith("https://") || value.startsWith("http://");
-}
-
-function collectCssUrls(value: unknown, urls: Set<string>) {
-  if (typeof value === "string") {
-    for (const match of value.matchAll(cssUrlPattern)) {
-      const url = match[2]?.trim();
-      if (url && isRemoteUrl(url)) {
-        urls.add(url);
-      }
-    }
-  } else if (Array.isArray(value)) {
-    for (const item of value) {
-      collectCssUrls(item, urls);
-    }
-  }
-}
-
-/** Every remote image URL a node tree references: `<img src>`, `backgroundImage`, `maskImage`. */
-function extractImageUrls(node: Node): string[] {
-  const urls = new Set<string>();
-
-  const visit = (current: Node) => {
-    const collectStyleUrls = (style: CSSProperties | undefined) => {
-      if (!style) {
-        return;
-      }
-
-      collectCssUrls(style.backgroundImage, urls);
-      collectCssUrls(style.maskImage, urls);
-    };
-
-    collectStyleUrls(current.style);
-    collectStyleUrls(current.preset);
-    collectCssUrls(current.tw, urls);
-
-    if (current.type === "image") {
-      if (typeof current.src === "string" && isRemoteUrl(current.src)) {
-        urls.add(current.src);
-      }
-      return;
-    }
-
-    if (current.type === "container") {
-      for (const child of current.children ?? []) {
-        visit(child);
-      }
-    }
-  };
-
-  visit(node);
-  return [...urls];
-}
-
-/** Fetches a URL's bytes, coalescing concurrent requests for the same URL through `cache`. A
- * rejected fetch is evicted so a later call can retry instead of replaying the failure. */
-function fetchImageData(
-  url: string,
-  options: FetchOptions,
-  fetchCache?: ImageFetchCache,
-): Promise<ArrayBuffer> {
-  const cached = fetchCache?.get(url);
-  if (cached) {
-    return cached;
-  }
-
-  const promise = fetchOk(url, options)
-    .then((response) => response.arrayBuffer())
-    .catch((error) => {
-      fetchCache?.delete(url);
-      throw error;
-    });
-
-  fetchCache?.set(url, promise);
-  return promise;
-}
-
-export type PrepareImagesOptions = FetchOptions & {
-  /** The node tree(s) whose remote images to fetch. */
-  node: Node | Node[];
-  /** Pre-fetched entries; their URLs are not re-fetched. */
-  sources?: ImageLoader[];
-  /** Single-flight byte cache shared across renders. */
-  fetchCache?: ImageFetchCache;
-  /** Throw on any fetch failure; if `false`, failed URLs are dropped. @default true */
-  throwOnError?: boolean;
-};
-
-/**
- * Collects every remote image a node tree references, fetches the ones not already in `sources`,
- * and returns them as `images` entries ready to hand to a renderer.
- */
-export async function prepareImages({
-  node,
-  sources = [],
-  fetchCache,
-  fetch,
-  timeout = defaultFetchTimeout,
-  signal,
-  throwOnError = true,
-}: PrepareImagesOptions): Promise<ImageLoader[]> {
-  const nodes = Array.isArray(node) ? node : [node];
-  const provided = new Map<string, ImageLoader>();
-
-  for (const image of sources) {
-    provided.set(image.src, image);
-  }
-
-  const urls = [...new Set(nodes.flatMap(extractImageUrls))].filter((url) => !provided.has(url));
-  const fetchOptions: FetchOptions = { fetch, timeout, signal };
-
-  const tasks = urls.map(async (src) => ({
-    src,
-    data: await fetchImageData(src, fetchOptions, fetchCache),
-  }));
-  const fetched = throwOnError
-    ? await Promise.all(tasks)
-    : (await Promise.allSettled(tasks))
-        .filter((result) => result.status === "fulfilled")
-        .map((result) => result.value);
-
-  return [...provided.values(), ...fetched];
-}
 
 /**
  * Images for a render: pre-fetched entries, or a group that also controls how remote images
@@ -288,7 +149,12 @@ async function collectImages(
     signal: options?.signal,
   });
 
-  return cache ? prepared.map((image) => ({ ...image, cache: image.cache ?? cache })) : prepared;
+  return cache
+    ? prepared.map((image) => ({
+        ...image,
+        cache: ("cache" in image ? image.cache : undefined) ?? cache,
+      }))
+    : prepared;
 }
 
 function mergeStylesheets(options: PipelineOptions | undefined, extra: string[]): string[] {
